+++ src/components/30-organisms/table/index.react.d.ts
@@ -2,6 +2,8 @@ import React from 'react';
 
 export interface AXATableProps {
   className?: string;
+  innerscroll?: number;
+  maxheight?: number;
 }
 
 declare function createAXATable(